WpEntity abstract base class and entity hydration
with the removal of setters, I needed to update the way I create objects. I decided to extract the createFromData() method to a contract and implement it in a base WpEntity abstract class. Now all of the WpEntity child classes can share the same code and create themselves from the wp API data